Channel: Added streams

+++ channel.ts
@@ -2,6 +2,7 @@ import type { Subscription } from "./events"
 import { ObjectMap } from "./map"
 
 type Function = (body: any) => any | Promise<any>
+type Stream = (body: any) => AsyncGenerator<any, void, any>
 interface Controller {
   response: (response: any) => void
   subscribe: (topic: string) => void
@@ -14,6 +15,7 @@ export class Channel {
   publish: (topic: string, body: any) => void = () => { }
   requests = new Map<number, PendingRequest>()
   private postApi = new ObjectMap<string, Function>
+  private streamApi = new ObjectMap<string, Stream>
   _events?: Map<string, Subscription>
   constructor() { }
   makeRequest(path: string, body: any | undefined, response: (response: Response) => void): Request {
@@ -23,6 +25,14 @@ export class Channel {
     }
     this.requests.set(id, pending)
     return { id, path, body }
+  }
+  makeStream(stream: string, body: any | undefined, response: (response: Response) => void): StreamRequest {
+    const id = this.id++
+    const pending: PendingRequest = {
+      request: { id, stream, body }, response
+    }
+    this.requests.set(id, pending)
+    return { id, stream, body }
   }
   makeSubscription(sub: string, body: any | undefined, response: (response: Response) => void): SubscriptionRequest {
     const id = this.id++
@@ -37,6 +47,10 @@ export class Channel {
   }
   post(path: string, request: Function) {
     this.postApi.set(path, request)
+    return this
+  }
+  stream(path: string, request: Stream) {
+    this.streamApi.set(path, request)
     return this
   }
   receive(some: any, controller: Controller) {
@@ -65,6 +79,12 @@ export class Channel {
       } catch (e) {
         if (id !== undefined) controller.response({ id, error: `${e}` })
       }
+    } else if (some.stream) {
+      const id: number | undefined = some.id
+      const api = this.streamApi.get(some.stream)
+      if (!api) throw 'api not found'
+      if (id === undefined) throw 'stream requires id'
+      this.streamRequest(id, controller, some.body, api)
     } else if (some.sub) {
       const id: number | undefined = some.id
       const subscription = this._events?.get(some.sub)
@@ -97,8 +117,21 @@ export class Channel {
       if (some.body) controller.event(some.topic, some.body)
     } else if (some.id !== undefined) {
       const request = this.requests.get(some.id)
-      this.requests.delete(some.id)
+      if (!request) return
+      if (!('stream' in request.request && !some.done)) {
+        this.requests.delete(some.id)
+      }
       if (request) request.response(some)
+    }
+  }
+  private async streamRequest(id: number, controller: Controller, body: any, stream: Stream) {
+    try {
+      for await (const value of stream(body)) {
+        controller.response({ id, body: value })
+      }
+      controller.response({ id, done: true })
+    } catch (e) {
+      controller.response({ id, error: `${e}` })
     }
   }
   events(events: Map<string, Subscription>) {
@@ -108,7 +141,7 @@ export class Channel {
 }
 
 interface PendingRequest {
-  request: Request | SubscriptionRequest
+  request: Request | SubscriptionRequest | StreamRequest
   response: (response: Response) => void
 }
 interface Request {
@@ -121,9 +154,15 @@ interface SubscriptionRequest {
   sub: string
   body?: any
 }
-interface Response {
+interface StreamRequest {
+  id: number
+  stream: string
+  body?: any
+}
+export interface Response {
   id: number
   topic?: string
   body?: any
   error?: any
+  done?: boolean
 }
